gorepogen: Use 1.x notation to target latest stable Go.
This way, there's no need to update .travis.yml files every time a new
version of Go comes out.

Use "master" instead of "tip" since that seems to be the new name for it.

Source: https://docs.travis-ci.com/user/languages/go#Specifying-a-Go-version-to-use

Huge thanks to @mvdan and @meatballhat for uncovering this!