wasmserve: create a simple command similar to gopherjs serve
It can be helpful for running a quick experiment with Go WebAssembly
in the browser, without needing to set things up in advance.

It supports quick iteration, because the Go command is recompiled on
each page refresh. It's also possible to reload just the WebAssembly
module by using the Alt+R shortcut.