wasmserve: use GOROOT of go command in PATH for wasm_exec.js
Previously, the GOROOT value was determined at the time the wasmserve
binary was compiled. This meant that if PATH was modified such that a
different go version is selected, the wrong copy of wasm_exec.js file
would be served.

This change fixes that by determining GOROOT at process start time.